pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "graduation-prj"
version = "0.1.0"
description = "基于 Qwen-VL 的安全生产管理智能体"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "langchain>=0.3.0",
  "transformers>=4.40.0",
  "torch>=2.1.0",
  "pillow>=10.0.0",
  "python-docx>=1.1.0",
  "pypdf>=4.0.0",
  "pymupdf>=1.24.0",
  "faiss-cpu>=1.8.0",
  "requests>=2.31.0",
  "matplotlib>=3.8.0",
  "numpy>=1.26.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
